+++ packages/react-components/react-headless-components-preview/stories/src/Avatar/index.stories.tsx
@@ -1,4 +1,4 @@
-import { Avatar } from '@fluentui/react-headless-components-preview';
+import { Avatar } from '@fluentui/react-headless-components-preview/avatar';
 
 import descriptionMd from './AvatarDescription.md';