+++ packages/react-components/react-headless-components-preview/stories/src/Switch/index.stories.tsx
@@ -1,4 +1,4 @@
-import { Switch } from '@fluentui/react-headless-components-preview';
+import { Switch } from '@fluentui/react-headless-components-preview/switch';
 
 import descriptionMd from './SwitchDescription.md';