+++ packages/react-components/react-headless-components-preview/stories/src/Toolbar/index.stories.tsx
@@ -5,7 +5,7 @@ import {
   ToolbarGroup,
   ToolbarRadioGroup,
   ToolbarToggleButton,
-} from '@fluentui/react-headless-components-preview';
+} from '@fluentui/react-headless-components-preview/toolbar';
 
 import descriptionMd from './ToolbarDescription.md';